+++ driver-core/CHANGELOG.rst
@@ -5,8 +5,10 @@ CHANGELOG
 ------
 
 - [api] The type of the map key returned by NoHostAvailable#getErrors has changed from
-        InetAddress to InetSocketAddress. Same for Initializer#getContactPoints return
-        and for AuthProvider#newAuthenticator.
+  InetAddress to InetSocketAddress. Same for Initializer#getContactPoints return and
+  for AuthProvider#newAuthenticator.
+- [api] The default load balacing policy is now DCAwareRoundRobinPolicy, and the local
+  datacenter is automatically picked based on the first connected node (JAVA-296)
 - [new] New optional AddressTranslater (JAVA-145)
 
 Merged from 1.0 branch: